Split playerinfodialog's mess into separate smaller files
