+++ src/playercard/playerinfodialog.py
@@ -1,173 +1,26 @@
 from __future__ import annotations
 
-from bisect import bisect_left
-from collections.abc import Sequence
-
-import numpy as np
-import pyqtgraph as pg
 from PyQt6.QtCore import QDateTime
-from PyQt6.QtCore import QObject
-from PyQt6.QtCore import QPointF
 from PyQt6.QtCore import Qt
-from PyQt6.QtCore import QThread
-from PyQt6.QtCore import pyqtSignal
-from PyQt6.QtGui import QColor
-from PyQt6.QtGui import QIcon
-from PyQt6.QtGui import QPixmap
-from PyQt6.QtWidgets import QListWidget
-from PyQt6.QtWidgets import QListWidgetItem
 from PyQt6.QtWidgets import QTableWidgetItem
-from PyQt6.QtWidgets import QTabWidget
-from pyqtgraph.graphicsItems.DateAxisItem import DateAxisItem
 
 import util
-from api.models.Avatar import Avatar
 from api.models.AvatarAssignment import AvatarAssignment
-from api.models.Leaderboard import Leaderboard
 from api.models.LeaderboardRating import LeaderboardRating
 from api.models.NameRecord import NameRecord
 from api.models.Player import Player
 from api.models.PlayerEvent import PlayerEvent
 from api.player_api import PlayerApiConnector
-from api.stats_api import LeaderboardApiConnector
 from api.stats_api import LeaderboardRatingApiConnector
-from api.stats_api import LeaderboardRatingJournalApiConnector
 from api.stats_api import LeagueSeasonScoreApiConnector
 from api.stats_api import PlayerEventApiAccessor
 from downloadManager import AvatarDownloader
-from downloadManager import DownloadRequest
-from model.rating import Rating
+from playercard.avatarhandler import AvatarHandler
 from playercard.leagueformatter import league_formatter_factory
+from playercard.ratingtabwidget import RatingTabWidgetController
 from playercard.statistics import StatsCharts
 
 FormClass, BaseClass = util.THEME.loadUiType("player_card/playercard.ui")
-
-Numeric = float | int
-
-
-class Crosshairs:
-    def __init__(self, plotwidget: pg.PlotWidget, series: LineSeries) -> None:
-        self.plotwidget = plotwidget
-        self.plotwidget.scene().sigMouseMoved.connect(self.update_lines_and_text)
-
-        self.series = series
-
-        pen = pg.mkPen("green", width=3)
-        self.xLine = pg.InfiniteLine(angle=90, pen=pen)
-        self.yLine = pg.InfiniteLine(angle=0, pen=pen)
-
-        self.plotwidget.addItem(self.xLine, ignoreBounds=True)
-        self.plotwidget.addItem(self.yLine, ignoreBounds=True)
-
-        color = QColor("black")
-        self.xText = pg.TextItem(color=color)
-        self.yText = pg.TextItem(color=color)
-
-        self.plotwidget.scene().addItem(self.xText)
-        self.plotwidget.scene().addItem(self.yText)
-
-        self.plotwidget.plotItem.getAxis("left").setWidth(40)
-        self._visible = True
-
-    def set_series(self, series: LineSeries) -> None:
-        self.series = series
-
-    def set_visible(self, visible: bool) -> None:
-        self.xLine.setVisible(visible)
-        self.yLine.setVisible(visible)
-        self.xText.setVisible(visible)
-        self.yText.setVisible(visible)
-
-    def display(self, *, seen: bool) -> None:
-        if not self._visible:
-            return
-        self.set_visible(seen)
-
-    def hide(self) -> None:
-        self.set_visible(False)
-
-    def show(self) -> None:
-        self.set_visible(True)
-
-    def change_visibility(self) -> None:
-        new_state = not self._visible
-        self.set_visible(new_state)
-        self._visible = new_state
-
-    def is_visible(self) -> bool:
-        return self._visible
-
-    def _closest_index(self, lst: Sequence[Numeric], value: Numeric) -> int:
-        pos = bisect_left(lst, value)
-        if pos == 0:
-            return pos
-        if pos == len(lst):
-            return pos - 1
-
-        before = lst[pos - 1]
-        after = lst[pos]
-        if after - value < value - before:
-            return pos
-        else:
-            return pos - 1
-
-    def map_to_data(self, pos: QPointF) -> QPointF:
-        view = self.plotwidget.plotItem.getViewBox()
-        value_pos = view.mapSceneToView(pos)
-        point_index = self._closest_index(self.series.x(), value_pos.x())
-        return self.series.point_at(point_index)
-
-    def get_xtext_pos(self, scene_point: QPointF) -> tuple[float, float]:
-        scene_width = self.plotwidget.sceneBoundingRect().width()
-        text_width = self.xText.boundingRect().width()
-        text_height = self.xText.boundingRect().height()
-        padding = 3
-
-        left_margin = self.plotwidget.plotItem.getAxis("left").width() - padding
-        right_margin = scene_width - text_width + padding
-
-        x = max(left_margin, scene_point.x() - text_width / 2)
-        x = min(x, right_margin)
-        y = self.plotwidget.sceneBoundingRect().bottom() - text_height + padding
-        return x, y
-
-    def get_ytext_pos(self, scene_point: QPointF) -> tuple[float, float]:
-        padding = 3
-        x = self.plotwidget.sceneBoundingRect().left() - padding
-        y = scene_point.y() - self.yText.boundingRect().height() / 2
-        return x, y
-
-    def update_lines_and_text(self, pos: QPointF) -> None:
-        if len(self.series.x()) == 0:
-            return
-
-        data_point = self.map_to_data(pos)
-        view = self.plotwidget.plotItem.getViewBox()
-        scene_point = view.mapViewToScene(data_point)
-
-        left_margin = self.plotwidget.plotItem.getAxis("left").width()
-        if scene_point.x() < left_margin or scene_point.y() < 0:
-            return
-
-        self.update_lines(pos, data_point)
-        self.update_text(scene_point, data_point)
-        self.show_at_pos(scene_point)
-
-    def update_text(self, scene_point: QPointF, data_point: QPointF) -> None:
-        date = QDateTime.fromSecsSinceEpoch(round(data_point.x())).toString("dd-MM-yyyy hh:mm")
-        self.xText.setHtml(f"<div style='background-color: #ffff00;'>{date}</div>")
-        self.xText.setPos(*self.get_xtext_pos(scene_point))
-        self.yText.setHtml(f"<div style='background-color: #ffff00;'>{data_point.y():.2f}</div>")
-        self.yText.setPos(*self.get_ytext_pos(scene_point))
-
-    def update_lines(self, pos: QPointF, data_point: QPointF) -> None:
-        if self.plotwidget.sceneBoundingRect().contains(pos):
-            self.xLine.setPos(data_point.x())
-            self.yLine.setPos(data_point.y())
-
-    def show_at_pos(self, pos: QPointF) -> None:
-        seen = self.plotwidget.sceneBoundingRect().contains(pos)
-        self.display(seen=seen)
 
 
 class PlayerInfoDialog(FormClass, BaseClass):
@@ -238,242 +91,3 @@ class PlayerInfoDialog(FormClass, BaseClass):
     def process_player_events(self, events: list[PlayerEvent]) -> None:
         for chartview in self.stats_charts.player_events_charts(events):
             self.statsChartsLayout.addWidget(chartview)
-
-
-class AvatarHandler:
-    def __init__(self, avatar_list: QListWidget, avatar_downloader: AvatarDownloader) -> None:
-        self.avatar_list = avatar_list
-        self.avatar_dler = avatar_downloader
-        self.requests = {}
-
-    def populate_avatars(self, avatar_assignments: list[AvatarAssignment] | None) -> None:
-        if avatar_assignments is None:
-            return
-
-        for assignment in avatar_assignments:
-            if self.avatar_dler.has_avatar(assignment.avatar.filename):
-                self._add_avatar(assignment.avatar)
-            else:
-                self._download_avatar(assignment.avatar)
-
-    def _prepare_avatar_dl_request(self, avatar: Avatar) -> DownloadRequest:
-        req = DownloadRequest()
-        req.done.connect(self._handle_avatar_download)
-        self.requests[avatar.url] = (req, avatar.tooltip)
-        return req
-
-    def _download_avatar(self, avatar: Avatar) -> None:
-        req = self._prepare_avatar_dl_request(avatar)
-        self.avatar_dler.download_avatar(avatar.url, req)
-
-    def _add_avatar(self, avatar: Avatar) -> None:
-        self._add_avatar_item(self.avatar_dler.get_avatar(avatar.filename), avatar.tooltip)
-
-    def _add_avatar_item(self, pixmap: QPixmap, description: str) -> None:
-        if pixmap.isNull():
-            icon = QIcon(util.THEME.pixmap("chat/avatar/avatar_blank.png").scaled(40, 20))
-        else:
-            icon = QIcon(pixmap.scaled(40, 20))
-        avatar_item = QListWidgetItem(icon, description)
-        self.avatar_list.addItem(avatar_item)
-
-    def _handle_avatar_download(self, url: str, pixmap: QPixmap) -> None:
-        _, tooltip = self.requests[url]
-        self._add_avatar_item(pixmap, tooltip)
-        del self.requests[url]
-
-
-class LineSeries:
-    def __init__(self, size: int = 0) -> None:
-        self._x: np.ndarray = np.zeros(size)
-        self._y: np.ndarray = np.zeros(size)
-
-    def x(self) -> np.ndarray:
-        return self._x
-
-    def y(self) -> np.ndarray:
-        return self._y
-
-    def set_point(self, index: int, point: QPointF) -> None:
-        self._x[index] = point.x()
-        self._y[index] = point.y()
-
-    def extend(self, series: LineSeries) -> None:
-        self._x = np.append(self._x, series.x())
-        self._y = np.append(self._y, series.y())
-
-    def point_at(self, index: int) -> QPointF:
-        return QPointF(self._x[index], self._y[index])
-
-
-class LineSeriesParser(QThread):
-    result_ready = pyqtSignal(LineSeries)
-
-    def __init__(self, unparsed_api_response: dict) -> None:
-        QThread.__init__(self)
-        self.data = unparsed_api_response
-
-    def run(self) -> None:
-        self.parse()
-
-    def parse(self) -> None:
-        journal = self.data["data"]
-        journal_leng = len(journal)
-
-        if journal_leng == 0:
-            self.result_ready.emit(LineSeries())
-            return
-
-        stats = self.data["included"]
-        stats_leng = len(stats)
-
-        series = LineSeries(stats_leng)
-
-        stats_index = journal_index = 0
-        while stats_index < stats_leng and journal_index < journal_leng:
-            if (
-                stats[stats_index]["id"]
-                != journal[journal_index]["relationships"]["gamePlayerStats"]["data"]["id"]
-            ):
-                journal_index += 1
-                continue
-
-            score_time_str = stats[stats_index]["attributes"]["scoreTime"]
-            score_time = QDateTime.fromString(score_time_str, Qt.DateFormat.ISODate)
-            # not creating additional objects (like Rating and QPointF)
-            # and not accessing their attributes in a loop will also give small
-            # improvement, but not quite noticeable (a few hundreds of a second
-            # per 10000 loop cycles -- ~10x less than API call deviation)
-            rating = Rating(
-                journal[journal_index]["attributes"]["meanAfter"],
-                journal[journal_index]["attributes"]["deviationAfter"],
-            )
-            point = QPointF(
-                score_time.toSecsSinceEpoch(),
-                rating.displayed(),
-            )
-            series.set_point(stats_index, point)
-            stats_index += 1
-            journal_index += 1
-
-        self.result_ready.emit(series)
-
-
-class RatingsPlotTab(QObject):
-    name_changed = pyqtSignal(int, str)
-
-    def __init__(
-            self,
-            index: int,
-            player_id: str,
-            leaderboard: Leaderboard,
-            plot: PlotController,
-    ) -> None:
-        super().__init__()
-        self.index = index
-        self.player_id = player_id
-        self.leaderboard = leaderboard
-        self.ratings_history_api = LeaderboardRatingJournalApiConnector()
-        self.ratings_history_api.ratings_ready.connect(self.process_rating_history)
-        self.plot = plot
-        self._loaded = False
-        self.workers = []
-
-    def __del__(self) -> None:
-        try:
-            self.clear_threads()
-        except RuntimeError:
-            pass
-
-    def enter(self) -> None:
-        if self._loaded:
-            return
-        self.name_changed.emit(self.index, "Loading...")
-        self.ratings_history_api.get_full_history(self.player_id, self.leaderboard.technical_name)
-
-    def clear_threads(self) -> None:
-        for worker in self.workers:
-            if worker.isRunning():
-                worker.quit()
-        self.workers.clear()
-
-    def finish(self) -> None:
-        self.clear_threads()
-        self.plot.draw_series()
-        self.name_changed.emit(self.index, self.leaderboard.pretty_name)
-
-    def process_rating_history(self, message: dict) -> None:
-        total_pages = message["meta"]["page"]["totalPages"]
-        current_page = message["meta"]["page"]["number"]
-        self._loaded = current_page == total_pages
-
-        worker = LineSeriesParser(message)
-        self.workers.append(worker)
-        worker.result_ready.connect(self.data_parsed)
-        worker.start()
-
-    def data_parsed(self, series: LineSeries) -> None:
-        self.plot.add_data(series)
-        if self._loaded:
-            self.finish()
-
-
-class RatingTabWidgetController:
-    def __init__(self, player_id: str, tab_widget: QTabWidget) -> None:
-        self.player_id = player_id
-        self.widget = tab_widget
-        self.widget.currentChanged.connect(self.on_tab_changed)
-
-        self.leaderboards_api = LeaderboardApiConnector()
-        self.leaderboards_api.data_ready.connect(self.populate_leaderboards)
-        self.tabs: dict[int, RatingsPlotTab] = {}
-
-    def run(self) -> None:
-        self.leaderboards_api.requestData()
-
-    def populate_leaderboards(self, message: dict[str, list[Leaderboard]]) -> None:
-        for index, leaderboard in enumerate(message["values"]):
-            widget = pg.PlotWidget()
-            tab = RatingsPlotTab(index, self.player_id, leaderboard, PlotController(widget))
-            tab.name_changed.connect(self.widget.setTabText)
-            self.tabs[index] = tab
-            self.widget.insertTab(index, widget, leaderboard.pretty_name)
-
-    def on_tab_changed(self, index: int) -> None:
-        self.tabs[index].enter()
-
-
-class PlotController:
-    def __init__(self, widget: pg.PlotWidget) -> None:
-        self.widget = widget
-        self.widget.setBackground("#202025")
-        self.widget.setAxisItems({"bottom": DateAxisItem()})
-        self.series = LineSeries()
-        self.crosshairs = Crosshairs(self.widget, self.series)
-        self.hide_irrelevant_plot_actions()
-        self.add_custom_menu_actions()
-
-    def clear(self) -> None:
-        self.widget.clear()
-
-    def draw_series(self) -> None:
-        self.widget.plot(self.series.x(), self.series.y(), pen=pg.mkPen("orange"))
-        self.widget.autoRange()
-
-    def add_data(self, series: LineSeries) -> None:
-        self.series.extend(series)
-
-    def hide_irrelevant_plot_actions(self) -> None:
-        for action in ("Transforms", "Downsample", "Average", "Alpha", "Points"):
-            self.widget.plotItem.setContextMenuActionVisible(action, visible=False)
-
-    def add_custom_menu_actions(self) -> None:
-        viewbox = self.widget.plotItem.getViewBox()
-        if viewbox is None:
-            return
-
-        menu = viewbox.getMenu(ev=None)
-        if menu is None:
-            return
-
-        menu.addAction("Show/Hide crosshair", self.crosshairs.change_visibility)
